# apps/api-gateway/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: AP Calculus Bot API
  description: API Gateway for AP Calculus AB/BC tutoring bot with verified answers
  version: 0.1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /review/:
    post:
      tags:
        - review
      description: Submit a case for human review
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                  minLength: 1
                  maxLength: 2000
                answer:
                  type: string
                  minLength: 1
                  maxLength: 10000
                examVariant:
                  type: string
                  enum:
                    - calc_ab
                    - calc_bc
                trustScore:
                  type: number
                  minimum: 0
                  maximum: 1
                confidence:
                  type: number
                  minimum: 0
                  maximum: 1
                sources:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum:
                          - canonical
                          - retrieval
                          - generated
                      id:
                        type: string
                      title:
                        type: string
                      snippet:
                        type: string
                      score:
                        type: number
                    required:
                      - type
                      - id
                    additionalProperties: false
                metadata:
                  type: object
                  properties:
                    examVariant:
                      type: string
                    topic:
                      type: string
                    subtopic:
                      type: string
                    difficulty:
                      type: string
                    processingTime:
                      type: number
                    retryCount:
                      type: number
                  additionalProperties: false
              required:
                - question
                - answer
                - examVariant
                - trustScore
                - confidence
              additionalProperties: false
        required: true
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    enum:
                      - pending
                      - approved
                      - rejected
                      - needs_revision
                  message:
                    type: string
                required:
                  - id
                  - status
                  - message
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
    get:
      tags:
        - review
      description: Get pending review cases (teacher only)
      parameters:
        - schema:
            type: string
            enum:
              - pending
              - approved
              - rejected
              - needs_revision
            default: pending
          in: query
          name: status
          required: false
        - schema:
            type: string
            enum:
              - calc_ab
              - calc_bc
          in: query
          name: examVariant
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          in: query
          name: limit
          required: false
        - schema:
            type: integer
            minimum: 0
            default: 0
          in: query
          name: offset
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  cases:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        question:
                          type: string
                        answer:
                          type: string
                        examVariant:
                          type: string
                        trustScore:
                          type: number
                        confidence:
                          type: number
                          nullable: true
                        sources:
                          type: array
                          items:
                            type: object
                            properties:
                              type:
                                type: string
                                enum:
                                  - canonical
                                  - retrieval
                                  - generated
                              id:
                                type: string
                              title:
                                type: string
                              snippet:
                                type: string
                              score:
                                type: number
                            required:
                              - type
                              - id
                            additionalProperties: false
                          default: []
                        metadata:
                          type: object
                          properties:
                            examVariant:
                              type: string
                            topic:
                              type: string
                            subtopic:
                              type: string
                            difficulty:
                              type: string
                            processingTime:
                              type: number
                            retryCount:
                              type: number
                          additionalProperties: false
                        status:
                          type: string
                          enum:
                            - pending
                            - approved
                            - rejected
                            - needs_revision
                        created_at:
                          type: string
                        updated_at:
                          type: string
                      required:
                        - id
                        - question
                        - answer
                        - examVariant
                        - trustScore
                        - status
                        - created_at
                        - updated_at
                      additionalProperties: false
                  pagination:
                    type: object
                    properties:
                      total:
                        type: number
                      limit:
                        type: number
                      offset:
                        type: number
                      hasMore:
                        type: boolean
                    required:
                      - total
                      - limit
                      - offset
                      - hasMore
                    additionalProperties: false
                required:
                  - cases
                  - pagination
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
  /review/resolve:
    post:
      tags:
        - review
      description: Resolve a review case (teacher only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                caseId:
                  type: string
                action:
                  type: string
                  enum:
                    - approve
                    - reject
                    - request_revision
                feedback:
                  type: string
                  maxLength: 1000
                correctedAnswer:
                  type: string
                  maxLength: 10000
                tags:
                  type: array
                  items:
                    type: string
              required:
                - caseId
                - action
              additionalProperties: false
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    enum:
                      - pending
                      - approved
                      - rejected
                      - needs_revision
                  message:
                    type: string
                required:
                  - id
                  - status
                  - message
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        description: Human readable error message
                      statusCode:
                        type: number
                        description: HTTP status code associated with the error
                      code:
                        type: string
                        description: Application-specific error code
                      validation:
                        description: Validation errors when applicable
                      details:
                        description: Additional error context
                    required:
                      - message
                      - statusCode
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
...
servers:
  - url: http://localhost:3001
    description: Development server