# apps/api-gateway/src/routes/review.ts
import { FastifyPluginAsync } from 'fastify';
import { createLogger } from '@ap/shared/logger';
import { withSpan } from '@ap/shared/tracing';
import { supabaseService } from '@ap/shared/supabase';
import {
  reviewSubmitRequestSchema,
  reviewSubmitResponseSchema,
  reviewErrorResponseSchema,
  reviewQuerySchema,
  reviewCasesResponseSchema,
  reviewResolveRequestSchema,
  reviewResolveResponseSchema,
  reviewNotFoundResponseSchema,
  makeErrorResponse,
} from '../schemas';
import { z } from 'zod';
import { ensureErrorHandling } from '../utils/errorHandling';

const logger = createLogger('review-routes');

/**
 * Review case schema
 */
/*
const reviewCaseSchema = z.object({
  id: z.string(),
  question: z.string(),
  answer: z.string(),
  examVariant: z.enum(['calc_ab', 'calc_bc']),
  trustScore: z.number().min(0).max(1),
  confidence: z.number().min(0).max(1),
  sources: z.array(z.object({
    type: z.enum(['canonical', 'retrieval', 'generated']),
    id: z.string(),
    title: z.string().optional(),
    snippet: z.string().optional(),
    score: z.number().optional(),
  })),
  metadata: z.object({
    examVariant: z.string(),
    topic: z.string().optional(),
    subtopic: z.string().optional(),
    difficulty: z.string().optional(),
    processingTime: z.number(),
    retryCount: z.number(),
  }),
  status: z.enum(['pending', 'approved', 'rejected', 'needs_revision']),
  created_at: z.string(),
  updated_at: z.string(),
});
*/

type ReviewActionInput = z.infer<typeof reviewResolveRequestSchema>;

type ReviewCaseUpdate = {
  status: 'approved' | 'rejected' | 'needs_revision';
  updated_at: string;
  feedback?: string;
  corrected_answer?: string;
  tags?: string[];
};

/**
 * Review routes for HiTL (Human-in-the-Loop) functionality
 */
export const reviewRoutes: FastifyPluginAsync = async fastify => {
  ensureErrorHandling(fastify);

  // Submit a case for review
  fastify.post(
    '/',
    {
      schema: {
        description: 'Submit a case for human review',
        tags: ['review'],
        body: reviewSubmitRequestSchema,
        response: {
          201: reviewSubmitResponseSchema,
          400: reviewErrorResponseSchema,
          500: reviewErrorResponseSchema,
        },
      },
    },
    async (request, reply) => {
      const requestId = request.requestId ?? request.id;

      try {
        const requestData = reviewSubmitRequestSchema.parse(request.body);

        logger.info(
          {
            question: requestData.question,
            examVariant: requestData.examVariant,
            trustScore: requestData.trustScore,
            requestId,
          },
          'Submitting case for review'
        );

        // Create review case in database
        const reviewCase = await withSpan(
          'review_create_case',
          async () => {
            const { data, error } = await supabaseService
              .from('review_case')
              .insert({
                question: requestData.question,
                answer: requestData.answer,
                exam_variant: requestData.examVariant,
                trust_score: requestData.trustScore,
                confidence: requestData.confidence,
                sources: requestData.sources ?? [],
                metadata: requestData.metadata ?? {},
                status: 'pending',
              })
              .select()
              .single();

            if (error) {
              throw new Error(`Failed to create review case: ${error.message}`);
            }

            return data;
          },
          {
            attributes: {
              'review.exam_variant': requestData.examVariant,
              'review.trust_score': requestData.trustScore,
            },
          }
        );

        logger.info(
          {
            caseId: reviewCase.id,
            question: requestData.question,
            requestId,
          },
          'Review case created successfully'
        );

        void reply.status(201);
        return {
          id: reviewCase.id,
          status: 'pending',
          message: 'Case submitted for review',
        };
      } catch (error) {
        if (error instanceof z.ZodError) {
          logger.warn(
            {
              error: error.errors,
              body: request.body,
              requestId,
            },
            'Review case validation failed'
          );

          void reply.status(400);
          return makeErrorResponse(400, 'Invalid request parameters', {
            validation: error.errors,
          });
        }

        logger.error(
          {
            error: error instanceof Error ? error.message : 'Unknown error',
            body: request.body,
            requestId,
          },
          'Failed to create review case'
        );

        void reply.status(500);
        return makeErrorResponse(500, 'Internal server error');
      }
    }
  );

  // Get pending review cases
  fastify.get(
    '/',
    {
      schema: {
        description: 'Get pending review cases (teacher only)',
        tags: ['review'],
        querystring: reviewQuerySchema,
        response: {
          200: reviewCasesResponseSchema,
          400: reviewErrorResponseSchema,
          500: reviewErrorResponseSchema,
        },
      },
    },
    async (request, reply) => {
      const requestId = request.requestId ?? request.id;

      try {
        const queryParams = reviewQuerySchema.parse(request.query);

        logger.info(
          {
            status: queryParams.status,
            examVariant: queryParams.examVariant,
            limit: queryParams.limit,
            offset: queryParams.offset,
            requestId,
          },
          'Fetching review cases'
        );

        // Build query
        let query = supabaseService
          .from('review_case')
          .select('*', { count: 'exact' })
          .eq('status', queryParams.status)
          .order('created_at', { ascending: false })
          .range(queryParams.offset, queryParams.offset + queryParams.limit - 1);

        if (queryParams.examVariant) {
          query = query.eq('exam_variant', queryParams.examVariant);
        }

        const { data, error, count } = await query;

        if (error) {
          throw new Error(`Failed to fetch review cases: ${error.message}`);
        }

        const cases = (data ?? []).map(case_ => ({
          id: case_.id,
          question: case_.question,
          answer: case_.answer,
          examVariant: case_.exam_variant,
          trustScore: case_.trust_score,
          confidence: case_.confidence,
          sources: case_.sources ?? [],
          metadata: case_.metadata ?? {},
          status: case_.status,
          created_at: case_.created_at,
          updated_at: case_.updated_at,
        }));

        const total = count ?? 0;
        const hasMore = queryParams.offset + queryParams.limit < total;

        logger.info(
          {
            totalCases: total,
            returnedCases: cases.length,
            hasMore,
            requestId,
          },
          'Review cases fetched successfully'
        );

        return {
          cases,
          pagination: {
            total,
            limit: queryParams.limit,
            offset: queryParams.offset,
            hasMore,
          },
        };
      } catch (error) {
        if (error instanceof z.ZodError) {
          logger.warn(
            {
              error: error.errors,
              query: request.query,
              requestId,
            },
            'Review cases query validation failed'
          );

          void reply.status(400);
          return makeErrorResponse(400, 'Invalid query parameters', {
            validation: error.errors,
          });
        }

        logger.error(
          {
            error: error instanceof Error ? error.message : 'Unknown error',
            query: request.query,
            requestId,
          },
          'Failed to fetch review cases'
        );

        void reply.status(500);
        return makeErrorResponse(500, 'Internal server error');
      }
    }
  );

  // Resolve a review case
  fastify.post(
    '/resolve',
    {
      schema: {
        description: 'Resolve a review case (teacher only)',
        tags: ['review'],
        body: reviewResolveRequestSchema,
        response: {
          200: reviewResolveResponseSchema,
          400: reviewErrorResponseSchema,
          404: reviewNotFoundResponseSchema,
          500: reviewErrorResponseSchema,
        },
      },
    },
    async (request, reply) => {
      const requestId = request.requestId ?? request.id;
      let requestData: ReviewActionInput | undefined;

      try {
        const parsedRequest = reviewResolveRequestSchema.parse(request.body);
        requestData = parsedRequest;
        const actionRequest = parsedRequest;

        logger.info(
          {
            caseId: actionRequest.caseId,
            action: actionRequest.action,
            requestId,
          },
          'Resolving review case'
        );

        // Update review case status
        const updatedCase = await withSpan(
          'review_resolve_case',
          async () => {
            const updateData: ReviewCaseUpdate = {
              status:
                actionRequest.action === 'approve'
                  ? 'approved'
                  : actionRequest.action === 'reject'
                    ? 'rejected'
                    : 'needs_revision',
              updated_at: new Date().toISOString(),
            };

            if (actionRequest.feedback) {
              updateData.feedback = actionRequest.feedback;
            }

            if (actionRequest.correctedAnswer) {
              updateData.corrected_answer = actionRequest.correctedAnswer;
            }

            if (actionRequest.tags) {
              updateData.tags = actionRequest.tags;
            }

            const { data, error } = await supabaseService
              .from('review_case')
              .update(updateData)
              .eq('id', actionRequest.caseId)
              .select()
              .single();

            if (error) {
              throw new Error(`Failed to update review case: ${error.message}`);
            }

            if (!data) {
              throw new Error('Review case not found');
            }

            return data;
          },
          {
            attributes: {
              'review.case_id': actionRequest.caseId,
              'review.action': actionRequest.action,
            },
          }
        );

        // Create review action record
        await withSpan(
          'review_create_action',
          async () => {
            const { error } = await supabaseService.from('review_action').insert({
              case_id: actionRequest.caseId,
              action: actionRequest.action,
              feedback: actionRequest.feedback,
              corrected_answer: actionRequest.correctedAnswer,
              tags: actionRequest.tags ?? [],
              created_at: new Date().toISOString(),
            });

            if (error) {
              throw new Error(`Failed to create review action: ${error.message}`);
            }
          },
          {
            attributes: {
              'review.case_id': actionRequest.caseId,
              'review.action': actionRequest.action,
            },
          }
        );

        const derivedStatus =
          updatedCase.status ??
          (actionRequest.action === 'approve'
            ? 'approved'
            : actionRequest.action === 'reject'
              ? 'rejected'
              : 'needs_revision');

        const actionMessage =
          actionRequest.action === 'approve'
            ? 'approved'
            : actionRequest.action === 'reject'
              ? 'rejected'
              : 'marked for revision';

        logger.info(
          {
            caseId: actionRequest.caseId,
            action: actionRequest.action,
            newStatus: derivedStatus,
            requestId,
          },
          'Review case resolved successfully'
        );

        return {
          id: updatedCase.id ?? actionRequest.caseId,
          status: derivedStatus,
          message: `Case ${actionMessage} successfully`,
        };
      } catch (error) {
        if (error instanceof z.ZodError) {
          logger.warn(
            {
              error: error.errors,
              body: request.body,
              requestId,
            },
            'Review action validation failed'
          );

          void reply.status(400);
          return makeErrorResponse(400, 'Invalid request parameters', {
            validation: error.errors,
          });
        }

        if (error instanceof Error && error.message.includes('not found')) {
          logger.warn(
            {
              caseId: requestData?.caseId,
              requestId,
            },
            'Review case not found'
          );

          void reply.status(404);
          return makeErrorResponse(404, 'Review case not found', { code: 'REVIEW_NOT_FOUND' });
        }

        logger.error(
          {
            error: error instanceof Error ? error.message : 'Unknown error',
            body: request.body,
            requestId,
          },
          'Failed to resolve review case'
        );

        void reply.status(500);
        return makeErrorResponse(500, 'Internal server error');
      }
    }
  );

  logger.info('Review routes registered');

  await Promise.resolve();
};